Pull Hamlet into separate files
There are some hanger-onner files (mainly .scss) that don't have much at all to do
with this commit, but the crux of it is that I managed to pull out the template
functionality from the tightly-coupled Blaze .hs files into separate Hamlet files.

At this point in time, everything works as I'd like it to. This is a commit to revert
to when I inevitably fuck it all up.