Add new config file to give access to Static
I'm not sure if I'm missing something or if this is already a part of
Spock, but not having access to a static route is a huge pain in the ass
for deployment. I can't think of an easier way to get it there other
than adding a secondary Config.hs file on top of the blog.cfg file.

So I have redundancies. For right now it's better though...didn't feel
like lifting everything into an IO monad...eugh.